requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.23
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
